use canvas drawer for silkscreen apertures
