+++ src/textures/create-combined-board-textures.ts
@@ -6,7 +6,6 @@ import { calculateOutlineBounds } from "../utils/outline-bounds"
 import { createPadTextureForLayer } from "../utils/pad-texture"
 import { createPanelOutlineTextureForLayer } from "../utils/panel-outline-texture"
 import { createTraceTextureForLayer } from "../utils/trace-texture"
-import { clearPhysicalBoardApertures } from "./clear-physical-board-apertures"
 import { createCopperPourTextureForLayer } from "./create-copper-pour-texture-for-layer"
 import { createCopperTextTextureForLayer } from "./create-copper-text-texture-for-layer"
 import { createFabricationNoteTextureForLayer } from "./create-fabrication-note-texture-for-layer"
@@ -97,18 +96,14 @@ const applySoldermaskSurfaceFilter = (
 
 const createCombinedTexture = ({
   textures,
-  circuitJson,
   boardData,
   traceTextureResolution,
   layer,
-  soldermaskVisible,
 }: {
   textures: Array<THREE.CanvasTexture | null | undefined>
-  circuitJson: AnyCircuitElement[]
   boardData: PcbBoard
   traceTextureResolution: number
   layer: "top" | "bottom"
-  soldermaskVisible: boolean
 }): THREE.CanvasTexture | null => {
   const hasImage = textures.some((texture) => texture?.image)
   if (!hasImage) return null
@@ -133,16 +128,6 @@ const createCombinedTexture = ({
     const image = texture.image as HTMLCanvasElement
     ctx.drawImage(image, 0, 0, canvasWidth, canvasHeight)
   }
-
-  clearPhysicalBoardApertures({
-    ctx,
-    circuitJson,
-    bounds: boardOutlineBounds,
-    layer,
-    soldermaskVisible,
-    width: canvasWidth,
-    height: canvasHeight,
-  })
 
   applySoldermaskSurfaceFilter(ctx, canvasWidth, canvasHeight, {
     includeReflection: layer === "top",
@@ -302,6 +287,7 @@ export function createCombinedBoardTextures({
           boardData,
           silkscreenColor,
           traceTextureResolution,
+          soldermaskVisible: showMask,
         })
       : null
 
@@ -356,11 +342,9 @@ export function createCombinedBoardTextures({
         panelOutlineTexture,
         keepoutTexture,
       ],
-      circuitJson,
       boardData,
       traceTextureResolution,
       layer,
-      soldermaskVisible: showMask,
     })
     const maskedCopperTexture =
       showMask && showCopper
