+++ src/textures/create-silkscreen-texture-for-layer.ts
@@ -1,8 +1,8 @@
 import type { AnyCircuitElement, PcbBoard } from "circuit-json"
 import * as THREE from "three"
 import { TRACE_TEXTURE_RESOLUTION } from "../geoms/constants"
-import { getSoldermaskRenderBounds } from "./soldermask/soldermask-bounds"
 import { drawSilkscreenLayer } from "./silkscreen/silkscreen-drawing"
+import { getSoldermaskRenderBounds } from "./soldermask/soldermask-bounds"
 
 const isSilkscreenElement = (
   element: AnyCircuitElement,
@@ -14,23 +14,42 @@ const isSilkscreenElement = (
   return elementType.startsWith("pcb_silkscreen_")
 }
 
+const isOpenSurfaceAperture = (
+  element: AnyCircuitElement,
+  soldermaskVisible: boolean,
+) => {
+  if (element.type === "pcb_cutout") return true
+  if (element.type === "pcb_via") {
+    return !soldermaskVisible || element.is_tented !== true
+  }
+  if (element.type === "pcb_hole" || element.type === "pcb_plated_hole") {
+    return !soldermaskVisible || element.is_covered_with_solder_mask !== true
+  }
+  return false
+}
+
 export function createSilkscreenTextureForLayer({
   layer,
   circuitJson,
   boardData,
   traceTextureResolution = TRACE_TEXTURE_RESOLUTION,
   silkscreenColor = "rgb(255,255,255)",
+  soldermaskVisible = true,
 }: {
   layer: "top" | "bottom"
   circuitJson: AnyCircuitElement[]
   boardData: PcbBoard
   traceTextureResolution?: number
   silkscreenColor?: string
+  soldermaskVisible?: boolean
 }): THREE.CanvasTexture | null {
   const elements = circuitJson.filter((element) =>
     isSilkscreenElement(element, layer),
   )
   if (elements.length === 0) return null
+  const apertureElements = circuitJson.filter((element) =>
+    isOpenSurfaceAperture(element, soldermaskVisible),
+  )
 
   const bounds = getSoldermaskRenderBounds(circuitJson, boardData)
   const canvasWidth = Math.floor(bounds.width * traceTextureResolution)
@@ -53,6 +72,7 @@ export function createSilkscreenTextureForLayer({
     layer,
     bounds,
     elements,
+    apertureElements,
     silkscreenColor,
   })
